[Spark] Isolate DeltaTaskStatisticsTracker direct-use test
Move the tracker wire-up test into its own suite so column-mapping
StatsCollection variants no longer inherit it, matching the DBR export fix.
