+++ spark/src/test/scala/org/apache/spark/sql/delta/stats/StatsCollectionSuite.scala
@@ -96,81 +96,6 @@ class StatsCollectionSuite
     }
   }
 
-  test("DeltaTaskStatisticsTracker computes stats outside V1 write tracker") {
-    // The test setup builds a `StatisticsCollection` directly from a renamed in-memory DataFrame
-    // whose schema does not carry column-mapping metadata. That is fine for the default
-    // (NoMapping) suite, but the reference `statsCollector` SQL path resolves attributes
-    // differently in column-mapping subclasses, producing a stats-JSON mismatch with the
-    // executor-side tracker. The test is meant to verify wire-up between
-    // `DeltaWriteStatsUtils.getStatsColExpr` and `DeltaTaskStatisticsTracker`, not
-    // column-mapping correctness, so skip when column mapping is enabled.
-    assume(!columnMappingEnabled)
-    withTempDir { dir =>
-      val deltaLog = DeltaLog.forTable(spark, dir)
-
-      val data = Seq[java.lang.Integer](1, null, 3)
-        .toDF("value")
-        .coalesce(1)
-      data.write.format("delta").save(dir.getAbsolutePath)
-      val snapshot = deltaLog.update()
-
-      // Convert logical names to physical names before building the stats expression, mirroring
-      // the write path.
-      val dataRenamed = data.toDF(
-        data.columns.map(name => getPhysicalName(name, snapshot.schema)): _*)
-
-      val statsCollection = new StatisticsCollection {
-        override val spark = StatsCollectionSuite.this.spark
-        override def tableSchema: StructType = dataRenamed.schema
-        override def outputTableStatsSchema: StructType = dataRenamed.schema
-        override def outputAttributeSchema: StructType = dataRenamed.schema
-        override val statsColumnSpec = DeltaStatsColumnSpec(
-          None,
-          Some(
-            DeltaConfigs.DATA_SKIPPING_NUM_INDEXED_COLS.fromString(
-              DeltaConfigs.DATA_SKIPPING_NUM_INDEXED_COLS.defaultValue)
-          )
-        )
-        override def columnMappingMode: DeltaColumnMappingMode = snapshot.columnMappingMode
-        override val protocol: Protocol = snapshot.protocol
-        override def getDataSkippingStringPrefixLength: Int =
-          StatsCollectionUtils.getDataSkippingStringPrefixLength(spark, snapshot.metadata)
-      }
-
-      val (statsColExpr, resolvedStatsDataSchema) =
-        DeltaWriteStatsUtils.getStatsColExpr(
-          spark, dataRenamed.queryExecution.analyzed.output, statsCollection)
-      val filePath = new Path(dir.getAbsolutePath, "part-00000-test.parquet").toString
-      val tracker = new DeltaTaskStatisticsTracker(
-        resolvedStatsDataSchema,
-        statsColExpr,
-        new Path(dir.getAbsolutePath),
-        spark.sessionState.newHadoopConf())
-
-      tracker.newFile(filePath)
-      dataRenamed.queryExecution.toRdd.collect().foreach { row =>
-        tracker.newRow(filePath, row.copy())
-      }
-      val statsJson = tracker
-        .getFinalStats(taskCommitTime = 0L)
-        .asInstanceOf[DeltaFileStatistics]
-        .stats(new Path(filePath).getName)
-
-      val expectedStatsJson = dataRenamed
-        .select(statsCollection.statsCollector)
-        .select(to_json($"stats").as[String])
-        .collect()
-        .head
-      assert(statsJson === expectedStatsJson)
-
-      val stats = JsonUtils.fromJson[Map[String, Any]](statsJson)
-      assert(stats.contains("numRecords"))
-      assert(stats.contains("minValues"))
-      assert(stats.contains("maxValues"))
-      assert(stats.contains("nullCount"))
-    }
-  }
-
   test("gather stats") {
     withTempDir { dir =>
       val deltaLog = DeltaLog.forTable(spark, dir)
